requirements.txt
pybtex
latexcodec
PyYAML==6.0.2
python-frontmatter
python-dotenv
openai
